add authorization

+++ app/Http/Controllers/BlogController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Blog;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BlogController extends Controller
 {
@@ -16,6 +19,10 @@ class BlogController extends Controller
 
     public function show(Blog $blog)
     {
+        Gate::define('edit-blog', function(User $user=null, Blog $blog) {
+            return $blog->user->is($user); 
+        });
+
         return view('blogs.show', ['blog' => $blog]);
     }
 
@@ -36,7 +43,7 @@ class BlogController extends Controller
             'title' => request('title'),
             'slug' => request('slug'),
             'body' => request('body'),
-            'user_id' => 1
+            'user_id' => request()->user()->id,
         ]);
     
         return redirect('/blog');
@@ -44,12 +51,34 @@ class BlogController extends Controller
 
     public function edit(Blog $blog)
     {
+        Gate::define('edit-blog', function(User $user, Blog $blog) {
+            return $blog->user->is($user); 
+        });
+
+        if(Auth::guest()) {
+            return redirect('/login');
+        }
+
+        Gate::authorize('edit-blog', $blog);
+
+        // if($blog->user->isNot(Auth::user())) {
+        //     abort(403);
+        // }
+
         return view('blogs.edit', ['blog' => $blog]);
     }
 
     public function update(Blog $blog)
     {
+        if(Auth::guest()) {
+            return redirect('/login');
+        }
+
+        Gate::define('edit-blog', function(User $user, Blog $blog) {
+            return $blog->user->is($user); 
+        });
         //todo: authorize 
+        Gate::authorize('edit-blog', $blog);
 
         request()->validate([
             'title' => ['required', 'min:3', 'max:255'],
@@ -68,6 +97,16 @@ class BlogController extends Controller
 
     public function destroy(Blog $blog)
     {
+        Gate::define('edit-blog', function(User $user, Blog $blog) {
+            return $blog->user->is($user); 
+        });
+
+        if(Auth::guest()) {
+            return redirect('/login');
+        }
+        
+        Gate::authorize('edit-blog', $blog);
+        
         $blog->delete();
         return redirect('/blog');
     }
